+++ src/sdk/CHANGELOG.md
@@ -27,6 +27,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Currently only support getting `SensitivityLabelId` on `ISite` [jansenbe - Bert Jansen]
 - Admin library: Fixed issues with loading SharePoint Add-Ins [jansenbe - Bert Jansen]
 - Admin library: Extra parameters to configure the SharePoint Add-In and Azure ACS principal loading [jansenbe - Bert Jansen]
+- Changed `IList.DraftVersionVisibility` to use enum `DraftVisibilityType` instead of an `int` #1286 [jansenbe - Bert Jansen]
 
 ## [1.10.0]